updated division file

+++ src/main/java/com/calc/Division.java
@@ -1,7 +1,7 @@
 package com.calc;
 public class Division {
 	public void div(int a,int b) {
-		int c;
+		float c;
 		c=a / b;
 		System.out.println("Division of "+ a +" and "+ b +" is "+c);
 		}
